chore(TT-1178): create 1 shard and 0 replicas
Create new indices with 1 shard and 0 replicas.

We've been shrinking the graph index down to 1 shard in the past.
Creating new indices with 1 shard from the start saves us some effort.

We have limited space in the dev/QA Elasticsearch clusters, so we don't
want our indices to carry replicas with them. We'll bump up the number
of replicas in prod anyway.