chore(TT-1178): configure number of replicas/shards
Add --replicas and --shards options to the master CLI for configuring
newly created indices.

Default to one shard because we normally shrink the index down to one
shard before deploying into production.

Default to zero replicas because we're usually constrained for space
in dev/QA environments, and we normally configure many more replicas
in prod than we'd reasonably want in other environments.