DEV-236 add mappings export to esbuild-cli (#263)
* DEV-236 add mappings export to esbuild-cli

Add a `mappings` tool to the `esbuild-cli` command to export the
Elasticsearch mappings that esbuild generates from the GDC dictionary.
Produce mappings that can easily be copied into gdc-models, to keep it
in sync as new fields are added to the dictionary.

Read descriptions for fields with metadata under `common` in the
dictionary. Fix the descriptions in the mapping metadata for fields
like `created_datetime`, `state`, and so on. Don't check dictionary
properties for `term` since it doesn't exist anymore.

Make `ES_HOST` default to `localhost` so it's possible to run
`esbuild-cli` without setting it.